Refactor namespaces and classes for improved organization and clarity
