+++ src/ShapeCrawler/Constants.cs
@@ -1,4 +1,4 @@
-﻿namespace ShapeCrawler.Shared;
+﻿namespace ShapeCrawler;
 
 internal static class Constants
 {
@@ -7,6 +7,6 @@ internal static class Constants
     internal static decimal DefaultLeftAndRightMargin => 0.25m; // cm
 
     internal static decimal DefaultTopAndBottomMargin => 0.13m; // cm
-    
+
     internal static long DefaultRowHeightEmu => 370840L;
 }
