+++ src/ShapeCrawler/Media/IMediaShape.cs
