+++ src/ShapeCrawler/Presentations/Presentation.cs
@@ -7,7 +7,6 @@ using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Validation;
 using ShapeCrawler.Exceptions;
 using ShapeCrawler.Presentations;
-using ShapeCrawler.Shared;
 using A = DocumentFormat.OpenXml.Drawing;
 
 #if NETSTANDARD2_0
